Sort `os.listdir` results to make step discovery consistent.
The selftests for duplicated steps expect steps to be discovered in
an order like "step1.py" then "step2.py". However, `os.listdir`
lists the entries in a directory in an arbitrary order. This caused
issues in certain environments and not others (inconsistency).

Wrapping the call to `os.listdir` in `sorted` seems to fix this issue.

Addresses Issue #122 in behave/behave; URL:
https://github.com/behave/behave/issues/122#issuecomment-13989830